+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- [builder] Welcome FQDN properly set (goto.kiwix)
+
+### Changed
+
+- [builder] Using captive-portal:1.4 with offline fix
+- [builder] Using reverse-proxy:1.7 with welcome_fqdn
+
 ## [1.9.0] - 2024-02-09
 
 ### Added